+++ ChillaxScraps/Config.cs
@@ -8,6 +8,7 @@ namespace ChillaxScraps
     {
         public readonly ConfigEntry<int> masterSwordDmg;
         public readonly ConfigEntry<bool> evilBoink;
+        public readonly ConfigEntry<bool> ocarinaUniqueSongs;
         public readonly List<ConfigEntry<int>> entries = new List<ConfigEntry<int>>();
 
         public Config(ConfigFile cfg, List<Scrap> scraps)
@@ -15,6 +16,7 @@ namespace ChillaxScraps
             cfg.SaveOnConfigSet = false;
             masterSwordDmg = cfg.Bind("Items", "Master Sword damage", 4, "Only the chosen hero can grab this sword, so it's supposed to be strong.");
             evilBoink = cfg.Bind("Items", "Evil Boink", false, "Activate this to turn Boink into an evil bird, can have negative consequences.");
+            ocarinaUniqueSongs = cfg.Bind("Items", "Ocarina unique songs", false, "Activate this if you want every connected player to have a randomly selected Ocarina song assigned to them.");
             foreach (Scrap scrap in scraps)
             {
                 entries.Add(cfg.Bind("Spawn chance", scrap.asset.Split("/")[0], scrap.rarity));